move training scripts
