+++ benchmarking/training_scripts/mlp_on_fashion_mnist/mlp_on_fashion_mnist.py
