+++ benchmarking/training_scripts/mlp_on_fashion_mnist/requirements.txt
