+++ examples/launch_huggingface_classification.py
@@ -19,13 +19,14 @@ from pathlib import Path
 from sagemaker.huggingface import HuggingFace
 
 import syne_tune
-from benchmarks.definitions.distilbert_on_imdb import distilbert_imdb_default_params, distilbert_imdb_benchmark
 from syne_tune.backend.sagemaker_backend.sagemaker_backend import SagemakerBackend
 from syne_tune.backend.sagemaker_backend.sagemaker_utils import get_execution_role
 from syne_tune.optimizer.baselines import RandomSearch
 from syne_tune.tuner import Tuner
 from syne_tune.stopping_criterion import StoppingCriterion
 
+from benchmarking.training_scripts.distilbert_on_imdb.distilbert_on_imdb import \
+    distilbert_imdb_benchmark, distilbert_imdb_default_params
 
 if __name__ == '__main__':
     logging.getLogger().setLevel(logging.INFO)
